requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.13
pydantic==2.12.0
pydantic-settings==2.12.0
httpx==0.27.2
redis==5.1.1
rq==1.16.2
python-multipart==0.0.9
python-docx==1.1.2
pytest==8.4.2
